Modify Packaging Activity, Fragment
