+++ app/src/main/AndroidManifest.xml
@@ -9,18 +9,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.SignInActivity">
+        <activity android:name=".view.Activity.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.BaseActivity" />
+        <activity android:name=".view.Activity.BaseActivity" />
         <activity
-            android:name=".view.MyHomeActivity"
+            android:name=".view.Activity.MyHomeActivity"
             android:label="@string/title_activity_my_home" />
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".view.Activity.MainActivity">
         </activity>
     </application>
