+++ app/src/main/java/com/hyun/myapplication/view/Fragment/CalendarFragment.kt
@@ -1,4 +1,4 @@
-package com.hyun.myapplication.view
+package com.hyun.myapplication.view.Fragment
 
 
 import android.os.Bundle
@@ -18,13 +18,15 @@ private const val ARG_PARAM2 = "param2"
  * A simple [Fragment] subclass.
  *
  */
-class RecordFragment : Fragment() {
+class CalendarFragment : Fragment() {
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_record, container, false)
+        return inflater.inflate(R.layout.fragment_calendar, container, false)
     }
-
 
 
 }
