+++ app/src/main/java/com/hyun/myapplication/view/Fragment/PicturesFragment.kt
@@ -1,4 +1,4 @@
-package com.hyun.myapplication.view
+package com.hyun.myapplication.view.Fragment
 
 
 import android.os.Bundle
@@ -18,14 +18,14 @@ private const val ARG_PARAM2 = "param2"
  * A simple [Fragment] subclass.
  *
  */
-class FamilyFragment : Fragment() {
+class PicturesFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_family, container, false)
+        return inflater.inflate(R.layout.fragment_pictures, container, false)
     }
