+++ app/src/main/res/layout/activity_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.SignInActivity">
+    tools:context=".view.Activity.SignInActivity">
 
     <ImageView
         android:id="@+id/imageView4"
