+++ app/src/main/res/layout/fragment_calendar.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CalendarFragment">
+    tools:context=".view.Fragment.CalendarFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
