+++ app/src/main/res/layout/fragment_pictures.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.PicturesFragment">
+    tools:context=".view.Fragment.PicturesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
