+++ app/src/main/res/layout/fragment_record.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.RecordFragment">
+    tools:context=".view.Fragment.RecordFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
