+++ app/src/main/res/layout/fragment_todo.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.TodoFragment">
+    tools:context=".view.Fragment.TodoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
